watch console cli crds

+++ knative-operator/pkg/controller/knativeserving/consoleclidownload/consoleclidownload.go
@@ -10,7 +10,6 @@ import (
 	socommon "github.com/openshift-knative/serverless-operator/pkg/common"
 	consolev1 "github.com/openshift/api/console/v1"
 	routev1 "github.com/openshift/api/route/v1"
-	"go.uber.org/atomic"
 	"k8s.io/apimachinery/pkg/api/equality"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -20,12 +19,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-const knCLIDownload = "kn"
+const (
+	CLIDownloadCRDName = "consoleclidownloads.console.openshift.io"
+	knCLIDownload      = "kn"
+)
 
 var (
 	operatorNamespace = os.Getenv(common.NamespaceEnvKey)
 	log               = common.Log.WithName("consoleclidownload")
-	ConsoleInstalled  = atomic.NewBool(false)
 )
 
 // Apply installs kn ConsoleCLIDownload and its required resources when applicable
@@ -35,11 +36,11 @@ func Apply(instance *operatorv1beta1.KnativeServing, apiclient client.Client) er
 	if err != nil {
 		return err
 	}
-	// If console is not installed skip installing cdd resources
-	if !ConsoleInstalled.Load() {
-		return nil
+	// If console is installed install cdd resources
+	if common.ConsoleInstalled.Load() {
+		return reconcileKnConsoleCLIDownload(apiclient, instance, route)
 	}
-	return reconcileKnConsoleCLIDownload(apiclient, instance, route)
+	return nil
 }
 
 func reconcileKnConsoleCLIDownloadRoute(apiclient client.Client, instance *operatorv1beta1.KnativeServing) (*routev1.Route, error) {
@@ -115,9 +116,12 @@ func reconcileKnConsoleCLIDownload(apiclient client.Client, instance *operatorv1
 
 // Delete deletes kn ConsoleCLIDownload CO and respective deployment resources
 func Delete(instance *operatorv1beta1.KnativeServing, apiclient client.Client, scheme *runtime.Scheme) error {
-	log.Info("Deleting kn ConsoleCLIDownload CO")
-	if err := apiclient.Delete(context.TODO(), populateKnConsoleCLIDownload("", instance)); err != nil && !apierrors.IsNotFound(err) {
-		return fmt.Errorf("failed to delete kn ConsoleCLIDownload CO: %w", err)
+	// If console is not installed skip deleting cdd resources
+	if common.ConsoleInstalled.Load() {
+		log.Info("Deleting kn ConsoleCLIDownload CO")
+		if err := apiclient.Delete(context.TODO(), populateKnConsoleCLIDownload("", instance)); err != nil && !apierrors.IsNotFound(err) {
+			return fmt.Errorf("failed to delete kn ConsoleCLIDownload CO: %w", err)
+		}
 	}
 
 	log.Info("Deleting kn ConsoleCLIDownload Route")
